Handle HttpErrorResponses as messages in Sentry
