+++ src/app/modules/shared/services/logging/sentry.service.ts
@@ -9,6 +9,10 @@ import { LoggingService } from './logging.service';
 import { LogLevel } from './log-level.model';
 import { FileLoggingService } from './file-logging.service';
 import { makeFetchTransport } from '@sentry/browser';
+import { HttpErrorResponse } from '@angular/common/http';
+import { getHttpErrorResponseMessageAndCode } from 'src/app/core/helpers/http-error-response-helper';
+import { RegistrationDraftErrorCode } from 'src/app/core/services/draft/draft-model';
+import type { CaptureContext } from '@sentry/types';
 
 @Injectable({
   providedIn: 'root',
@@ -57,8 +61,8 @@ export class SentryService implements LoggingService {
     }
   }
 
-  log(message?: string, error?: Error, level?: LogLevel, tag?: string, optionalParams?: { [key: string]: any }) {
-    this.fileLoggingService.log(message, error, level, tag, optionalParams);
+  log(message?: string, error?: unknown, level?: LogLevel, tag?: string, optionalParams?: { [key: string]: any }) {
+    this.fileLoggingService.log(message, error as Error, level, tag, optionalParams);
     if (message && (level === LogLevel.Warning || level === LogLevel.Info || level === LogLevel.Error)) {
       const breadcrumb: Sentry.Breadcrumb = {
         category: tag,
@@ -73,7 +77,28 @@ export class SentryService implements LoggingService {
       Sentry.addBreadcrumb(breadcrumb);
     }
     if (error && level === LogLevel.Error) {
-      if (error) {
+      if (error instanceof HttpErrorResponse) {
+        // Angular HttpErrorResponses are not instance of Error and are just logged as
+        // "Object captured as exception with keys: error, headers, message, name, ok" in Sentry.
+        // See https://github.com/getsentry/sentry-javascript/issues/2292.
+        // Here we turn the HttpErrorResponse into useful info for Sentry.
+        const { message: errorMessage, code } = getHttpErrorResponseMessageAndCode(error);
+
+        const context: CaptureContext = {
+          level: code === RegistrationDraftErrorCode.NoNetworkOrTimedOut ? 'warning' : 'error',
+          extra: {
+            status: error.status,
+            statusText: error.statusText,
+            error: error.error,
+            message: error.message,
+            url: error.url,
+          },
+        };
+
+        Sentry.captureMessage(errorMessage, context);
+      } else {
+        // Assume instance of error, we can add additional if checks if we see that we more custom error objects are
+        // thrown and we are getting more "Object captured as exception with keys" events in Sentry
         Sentry.captureException(error);
       }
     }
